requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
numpy==2.1.3
openai==1.52.0
python-dotenv==1.0.1

